\name{survo.sounds}
\alias{survo.sounds}
\alias{survo.sounds}
\docType{data}
\title{
Sounds for Survo R
}
\description{
Default samples for Survo R in audioSample format.}
\usage{data(survo.sounds)}
\format{
  The format is:
List of 72
 $ default
 $ APPLAUSE
 $ bilj
 $ biljs
 $ CAMERA
 $ CARBRAKE
 $ CASHREG
 $ CHIMES
 $ CHK_ERR
 $ CLAP
 $ Cuculus.canorus
 $ DRIVEBY
 $ DRUMROLL
 $ EXPLODE
 $ Fringilla.coelebs
 $ GLASS
 $ GUNSHOT
 $ kahdeks
 $ kaksi
 $ kmmenen
 $ kolme
 $ kuusi
 $ kymm
 $ KYMMENEN
 $ LASER
 $ miinus
 $ milj
 $ miljs
 $ milrd
 $ milrds
 $ nelja
 $ newalert
 $ nolla
 $ offspread
 $ pilkku
 $ plus
 $ PROJCTOR
 $ PROTECT
 $ RICOCHET
 $ ringout
 $ sata
 $ sataa
 $ seitsem
 $ SM1
 $ SM2
 $ SM3
 $ SM4
 $ SM5
 $ SM6
 $ SM7
 $ SM8
 $ SP1
 $ SP2
 $ SP3
 $ SP4
 $ SP5
 $ SP6
 $ SP7
 $ SP8
 $ SUR_ACT
 $ SUR_ML
 $ SUR_MR
 $ toista
 $ tuhat
 $ tuhatta
 $ TYPE
 $ TYPE2
 $ viisi
 $ WHOOSH
 $ XXX1
 $ yhdeks
 $ yksi
}
\details{
Samples in audioSample format (.WAV-files loaded with load.wave()-function from
the audio-package). 
}
\source{
Samples from the Survo MM distribution.
}
\keyword{datasets}
